phase_8_guardrails/requirements.txt
pyyaml>=6.0
groq>=0.9.0          # optional — only needed when use_llm: true
